+++ packages/components-react/src/Alert.tsx
@@ -15,13 +15,14 @@ export interface AlertProps extends UtrechtAlertProps {
   type?: UtrechtAlertType;
   icon?: RHCIconID;
 }
+
 const iconMap: Record<UtrechtAlertType, RHCIconID> = {
   info: 'info-circle',
   ok: 'circle-check',
   warning: 'let-op',
   error: 'alert-circle',
 };
-export const Alert = ({ type, icon, className, ...restProps }: PropsWithChildren<AlertProps>) => {
+export const Alert = ({ type, icon, ...restProps }: PropsWithChildren<AlertProps>) => {
   const rhcIcon = icon ? <Icon icon={icon} /> : type ? <Icon icon={iconMap[type]} /> : undefined;
   return <UtrechtAlert className={clsx('rhc-alert', className)} {...restProps} icon={rhcIcon} type={type} />;
 };