+++ packages/components-react/src/Blockquote.tsx
@@ -2,6 +2,13 @@
  * @license EUPL-1.2
  * Copyright (c) 2025 Community for NL Design System
  */
+import { BlockquoteProps, Blockquote as UtrechtBlockquote } from '@utrecht/component-library-react';
+import clsx from 'clsx';
 
-export { Blockquote } from '@utrecht/component-library-react';
 export type { BlockquoteProps } from '@utrecht/component-library-react';
+
+export const Blockquote = ({ className, ...restProps }: BlockquoteProps) => {
+  return <UtrechtBlockquote className={clsx('rhc-blockquote', className)} {...restProps} />;
+};
+
+Blockquote.displayName = 'Blockquote';