+++ packages/components-react/src/BreadcrumbNav.tsx
@@ -4,44 +4,45 @@
  */
 
 import {
-  BreadcrumbNav,
-  type BreadcrumbNavProps,
   BreadcrumbNavSeparator,
   type BreadcrumbNavSeparatorProps,
+  BreadcrumbNav as UtrechtBreadcrumbNav,
   BreadcrumbNavLink as UtrechtBreadcrumbNavLink,
   type BreadcrumbNavLinkProps as UtrechtBreadcrumbNavLinkProps,
+  type BreadcrumbNavProps as UtrechtBreadcrumbNavProps,
 } from '@utrecht/component-library-react';
 import clsx from 'clsx';
-import { PropsWithChildren, Ref } from 'react';
+import { Ref } from 'react';
 
-export { BreadcrumbNav, type BreadcrumbNavProps, BreadcrumbNavSeparator, type BreadcrumbNavSeparatorProps };
+type BreadcrumbNavProps = UtrechtBreadcrumbNavProps & {
+  ref?: Ref<HTMLOListElement>;
+};
+
+export { type BreadcrumbNavProps, BreadcrumbNavSeparator, type BreadcrumbNavSeparatorProps };
 
 export interface BreadcrumbNavLinkProps extends UtrechtBreadcrumbNavLinkProps {
   active?: boolean;
   ref?: Ref<HTMLAnchorElement>;
 }
 
-export const BreadcrumbNavLink = ({
-  ref,
-  children,
-  className,
-  href,
-  active,
-  ...restProps
-}: PropsWithChildren<BreadcrumbNavLinkProps>) => {
-  return (
-    <UtrechtBreadcrumbNavLink
-      href={href}
-      className={clsx({
-        'utrecht-breadcrumb-nav__link--active': active,
-        className,
-      })}
-      {...restProps}
-      ref={ref}
-    >
-      {children}
-    </UtrechtBreadcrumbNavLink>
-  );
-};
+export const BreadcrumbNavLink = ({ children, className, active, current, ...restProps }: BreadcrumbNavLinkProps) => (
+  <UtrechtBreadcrumbNavLink
+    current={current}
+    {...restProps}
+    className={clsx({
+      'rhc-breadcrumb-nav__link--active': active,
+      'rhc-breadcrumb-nav__link--current': current,
+      className,
+    })}
+  >
+    {children}
+  </UtrechtBreadcrumbNavLink>
+);
 
 BreadcrumbNavLink.displayName = 'BreadcrumbNavLink';
+
+export const BreadcrumbNav = ({ className, ...restProps }: BreadcrumbNavProps) => (
+  <UtrechtBreadcrumbNav className={clsx('rhc-breadcrumb-nav', className)} {...restProps} />
+);
+
+BreadcrumbNav.displayName = 'BreadcrumbNav';